Add config.h generation based on XML properties and jinja2
Jinja2 is a python template engine used to generate document using
advanced features such as loop and condition.

This tool is used to generate the Config class based only on the property
name, type and default value.

attribute name, description name and and ini name are automatically generated
from a space separated string.