Rework UIC compilation
Auto UIC is now disabled as it intented to generate a different UIC header for
each header in different directory.

To avoid this issue and as cmake UIC target variable doesn't affect anything,
two target uic and uic-settings are created and explicitly call qt5_wrap_uic
on UIC files and add thes files as dependencies.

Target view and view-settings dependes on the appropriate uic targets.