Merge points used to merge polylines
Previously when two polyline were detected near enough to be merged
the two point used in the test were left as it. This introduced later
error because the final polyline was containing bulges that do not
actually connect.

To avoid this issue, when two polylines are merged in the assembler
a middle point is computed between their merging points and used as
new point for the end of the first polyline and for the start of the
second polyline.