Fix selection and use std::set for tracking selected paths
Path::setSelect was triggering an event when the state was the same,
it must do the opposite.

Model::Task now use a std::set for tracking selected paths as this
container is safer in case of double selection/deselection.

Model::PathGroupSetting could not rely on std::vector for selected
paths, it use only Task::forEachSelectedPath.