feat: manage cutting direction to reduce burrs
Experience showed that cutting direction matters in burr reduction process.
Cutting in direction so that mill flutes in contact with current path are always
pulling material instead of pushing reduce a lot burrs.

The cutting direction is based on the direction of the offset the user wanted,
if the user offseted right a path then the path should be followed in opposite direction
(= cutting backward), in left offset nothing change.

This information about the offset direction is stored in OffsettedPath class and accessed
through Path class used where cuttingDirection getter is used in gcode Exporter.