feat: Introduce an optimizer for path cut order respecting group order.
This optimizer take as input a list of node sorted by group and compute an optimum
order with minimal distance when passing through all nodes.

By definition all nodes of the same groups are next to each other and the
nodes of the group i are selected before nodes of group i + 1.

In a concrete case, these groups will mean paths of the relativly same length.